backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.38
psycopg[binary]==3.2.6
pydantic[email]==2.10.6
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
stripe==11.6.0
reportlab==4.2.5
python-multipart==0.0.20
